+++ tomcat/src/main/java/com/techcourse/controller/RegisterController.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.techcourse.controller;
 
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
